feat defines example section for messaging container
The convention so far involves referencing other fields by
bounding the name with $. It's adequate and easy enough
to type. It'll be clearer whether this is fine when the file
is finished. The backtick escaping is a little ugly, but
that's in comparison to a more dsl style, rather than
another parser existant format (ex json, toml). The list/map
references are nice (-e -p).