feat 191104_20 start of representing docker container
Fills in mongo section of example eno config. Introduces ^name^
to indicate this should only resolve within the container
section or at runtime, rather than across sections. Removes the
docker recipe subsection, as it was redundant with the recipe
field. If need be, we can offset it again.

Adds DockerRunArg which can represent a single string or a map
of values. In this fashion it can represent the detach flag
and a list of volume mappings. This is in preference to using
separate, one key maps for each volume argument for that container.