feat i700101 represents container ports,volumes
This checkin is also to vacate the location, rather than
a complete task. This allows for rudimentary container
prep. Starts emitting a bash string, albiet to stdout.
Adds a dependence on org.json to parse the tink password.
Uses Tink to generate a password, if requested.

JgJiet fills the JgjTask's relevant containers when added.
This requires adding containers before the task.
BaseContainer has a mapping of ports, volumes, and env
variables, rather than a single internal port.